Velocidade linear para float

+++ app/models.py
@@ -58,7 +58,7 @@ class Teste(models.Model):
 	
 class TesteVelocidadeFixa(Teste):
 
-	testeVF_velocidade = models.IntegerField(null = True, verbose_name='Velocidade do Motor',default=5,
+	testeVF_velocidade = models.FloatField(null = True, verbose_name='Velocidade do Motor',default=5,
         validators=[
             MaxValueValidator(100),
             MinValueValidator(1)])
